requirements.txt
aioflask==0.4.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
asyncio==3.4.3
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
discord.py==2.4.0
Flask==3.0.3
frozenlist==1.4.1
greenlet==3.1.0
greenletio==0.11.0
h11==0.14.0
idna==3.9
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.1.0
numpy==2.1.1
pillow==10.4.0
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.3
uvicorn==0.30.6
Werkzeug==3.0.4
yarl==1.11.1
aiofiles==24.1.0